+++ src/utils/mouse.h
@@ -1,6 +1,6 @@
 #pragma once
-#include <SDL_mouse.h>
+#include <SDL_mouse.h>  // IWYU pragma: export
 #include <glm/vec2.hpp> // IWYU pragma: export
 
-glm::vec2
+glm::ivec2
 mouse_position();
